Refactoring routes

+++ app/Http/Controllers/AnimalController.php
@@ -2,14 +2,64 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Animals;
-use Illuminate\Http\Request;
 use Storage;
+use Inertia\Inertia;
+use App\Enums\Gender;
+use App\Models\Breeds;
+use App\Models\Animals;
+use App\Enums\AnimalStatus;
+use Illuminate\Http\Request;
 use App\Concerns\HandleImageUpload;
 
 class AnimalController extends Controller
 {
     use HandleImageUpload;
+
+    public function show()
+    {
+        return Inertia::render('animals/animals', [
+            'success' => session('success'),
+            'animals' => Animals::with('breed')
+                ->inRandomOrder()
+                ->paginate(10)]);
+        {/*->through(function ($animal) {
+                $animal->photo_url = Storage::disk('s3')->url($animal->photo);
+                return $animal;
+            })]);*/}
+    }
+
+    public function create()
+    {
+        $user = auth()->user();
+
+        if (!$user->organization) {
+            return redirect()->route('animals')->with('access', 'Vous devez appartenir à une organisation pour ajouter un animal.');
+        }
+
+        return Inertia::render('animals/create', [
+            'organization' => $user->organization,
+            'statuses' => AnimalStatus::cases(),
+            'breeds' => Breeds::all(),
+            'gender' => Gender::cases()
+        ]);
+    }
+
+    public function edit(Animals $animal)
+    {
+        $user = auth()->user();
+
+        if (!$user->organization) {
+            return redirect()->route('animals.show', $animal)->with('access', 'Vous devez appartenir à une organisation pour modifier un animal.');
+        }
+
+        return Inertia::render('animals/edit', [
+            'animal' => $animal,
+            'organization' => $user->organization,
+            'statuses' => AnimalStatus::cases(),
+            'breeds' => Breeds::all(),
+            'gender' => Gender::cases()
+        ]);
+    }
 
     public function store(Request $request)
     {
@@ -23,7 +73,7 @@ class AnimalController extends Controller
             'gender' => 'required|in:Mâle,Femelle',
             'adoption_status' => 'required|in:Disponible,En attente,Adopté',
             'photo' => 'required|image|mimes:jpeg,png,jpg,svg,webp|max:1024',
-            'description' => 'nullable|string|max:255'
+            'description' => 'nullable|string|max:2000'
         ]);
 
         if ($request->hasFile('photo')) {
@@ -37,6 +87,11 @@ class AnimalController extends Controller
 
     public function destroy(Animals $animal)
     {
+        $user = auth()->user();
+
+        if (!$user->organization) {
+            return redirect()->route('animals.show', $animal)->with('access', 'Vous devez appartenir à une organisation pour supprimer un animal.');
+        }
         if ($animal->photo) {
             Storage::disk('public')->delete($animal->photo);
         }
@@ -57,8 +112,8 @@ class AnimalController extends Controller
             'breed_id' => 'required|exists:breeds,id',
             'gender' => 'required|in:Mâle,Femelle',
             'adoption_status' => 'required|in:Disponible,En attente,Adopté',
-            'photo' => 'required|image|mimes:jpeg,png,jpg,svg,webp|max:1024',
-            'description' => 'nullable|string|max:255'
+            //'photo' => 'nullable|image|mimes:jpeg,png,jpg,svg,webp|max:1024',
+            'description' => 'nullable|string|max:2000'
         ]);
 
         if ($request->hasFile('photo')) {
